Coordinates to doubles

+++ src/Coordinate.java
@@ -1,24 +1,24 @@
 public class Coordinate {
-    private int x;
-    private int y;
+    private double x;
+    private double y;
 
-    public Coordinate (int x, int y)  {
+    public Coordinate (double x, double y)  {
         this.x = x;
         this.y = y;
     }
 
-    public int getXCoordinate() {
+    public double getXCoordinate() {
         return x;
     }
-    public int getYCoordinate() {
+    public double getYCoordinate() {
         return y;
     }
 
-    public void setX(int x) {
+    public void setX(double x) {
         this.x = x;
     }
 
-    public void setY(int y) {
+    public void setY(double y) {
         this.y = y;
     }
 }
